refactor : outputview 수정
